Add files via upload
The Pico 2W acts as the master in this configuration. Core 0 performs communication tasks with the slave, while Core 1 performs all http duties, including making a web server and generating an interactive web page. The master is connected by four wires to the slave, using standard GPIO pins (see lines 8-12 in spi_master.h). Core 0 receives three bytes of data from the slave. Byte 2 is an eight bit buffer for the LED states: 00000001 = green; 00000010 = yellow; 00000100 = red, and every combination thereof. Bytes 0 and 1 hold the ADC value, which was sent from the slave.

Core 0 delivers these bytes to the web interface, which is managed on Core 1. The web page has output fields for raw data, temperature in Fahrenheit, and the LED states. It will turn on graphic representations of the three LEDs, R,Y,G. Below the LED graphics are buttons which toggle their respective LED on or off. The master sends the slave only one byte of data, which I call the accessory byte, and which contains the value of a LED that has been toggled, if any. The button events are "one shot," which means after sending the value in one byte one time the sent byte will return to a value of 00000000. It remains at zero until a button is clicked.

I've tried several different master/slave programs for the Pico series online, but only Mr. Stoops' code worked reliably and robustly, and was easily extensible. The master controls the pacing, and in the original example the cycles were set to 10 seconds, but I brought it down to one second so that it turns the LEDs on and off faster.